backend/requirements.txt
Django==5.1.5
django-auditlog==2.3.0
django-guardian==2.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
whitenoise[brotli]==6.7.0
gunicorn==22.0.0